Logo tidying.
